Expose Components, rearrange types
