+++ src/components/CioPlp/CioPlp.tsx
@@ -1,7 +1,9 @@
 import React from 'react';
-import { CioPlpProps, IncludeRenderProps, PlpContextValue } from '../../types';
+import { IncludeRenderProps, PlpContextValue, CioPlpProviderProps } from '../../types';
 import CioPlpProvider from './CioPlpProvider';
 import CioPlpGrid from '../CioPlpGrid';
+
+export type CioPlpProps = CioPlpProviderProps;
 
 // Wrapper component for CioPlpProvider with default Markup
 export default function CioPlp(props: IncludeRenderProps<CioPlpProps, PlpContextValue>) {
