+++ src/components/CioPlp/index.ts
@@ -1,4 +1,5 @@
 import CioPlp from './CioPlp';
 
 export { default as CioPlpProvider } from './CioPlpProvider';
+export * from './CioPlp';
 export default CioPlp;
