+++ src/index.ts
@@ -1,4 +1,5 @@
 // Hooks
+export { useCioPlpContext } from './hooks/useCioPlpContext';
 export { default as useCioPlp } from './hooks/useCioPlp';
 export { default as useSearchResults } from './hooks/useSearchResults';
 export { default as useBrowseResults } from './hooks/useBrowseResults';
@@ -7,9 +8,21 @@ export { default as useSort } from './hooks/useSort';
 export * from './hooks/callbacks';
 
 // Components
-export { default as CioPlp } from './components/CioPlp';
-export { default as ProductCard, ProductCardProps } from './components/ProductCard';
+export { default as CioPlp, CioPlpProvider } from './components/CioPlp';
+export { default as CioPlpGrid, ZeroResults } from './components/CioPlpGrid';
+export { default as ProductCard } from './components/ProductCard';
 export { default as Sort } from './components/Sort';
+export { default as Pagination } from './components/Pagination';
+export { default as Filters } from './components/Filters';
+export { default as ProductSwatch } from './components/ProductSwatch';
+export { default as Spinner } from './components/Spinner';
 
-export { useCioPlpContext } from './hooks/useCioPlpContext';
+// Types
+export type { CioPlpProps } from './components/CioPlp';
+export type { CioPlpGridProps, CioPlpGridWithRenderProps } from './components/CioPlpGrid';
+export type { ProductCardProps } from './components/ProductCard';
+export type { ProductSwatchProps } from './components/ProductSwatch';
+export type { SortProps, SortWithRenderProps } from './components/Sort';
+export type { PaginationProps, PaginationWithRenderProps } from './components/Pagination';
+export type { FiltersProps, FiltersWithRenderProps } from './components/Filters';
 export * from './types';
